rename diff_iter.rs to diff.rs
