+++ ratatui-core/src/buffer/diff.rs
